requirements.txt
pyrofork
aiosqlite
python-dotenv
dateparser
apscheduler
psutil
dnspython
